+++ Sources/KeyboardKit/Input/InputSetProviderBased.swift
@@ -13,6 +13,10 @@ import Foundation
  This protocol is implemented by services that may depend on
  an ``InputSetProvider`` and must be reconfigured when a new
  input set provider is being used.
+
+ > Note: This will no longer be needed when the library uses
+ keyboard layout providers that use their own providers. The
+ context provider can then be removed and this as well.
  */
 public protocol InputSetProviderBased {