+++ Sources/KeyboardKit/Keyboard/KeyboardType.swift
@@ -21,8 +21,8 @@ public enum KeyboardType: Codable, Equatable, Identifiable {
      `.alphabetic` represents keyboards that have alphabetic
      input keys for the current locale.
      
-     This can be implemented with a `SystemKeyboard` but you
-     can create custom alphabetic keyboards as well.
+     This type can be created with a ``SystemKeyboard``, but
+     you can create custom alphabetic keyboards as well.
      */
     case alphabetic(KeyboardCasing)
     
@@ -30,8 +30,8 @@ public enum KeyboardType: Codable, Equatable, Identifiable {
      `.numeric` represents keyboards that have numeric input
      keys for the current locale.
      
-     This can be implemented with a `SystemKeyboard` but you
-     can create custom numeric keyboards as well.
+     This type can be created with a ``SystemKeyboard``, but
+     you can create custom numeric keyboards as well.
      */
     case numeric
     
@@ -39,8 +39,8 @@ public enum KeyboardType: Codable, Equatable, Identifiable {
      `.symbolic` represents keyboards that have symbol input
      keys for the current locale.
      
-     This can be implemented with a `SystemKeyboard` but you
-     can create custom symbolic keyboards as well.
+     This type can be created with a ``SystemKeyboard``, but
+     you can create custom symbolic keyboards as well.
      */
     case symbolic
     
@@ -48,9 +48,8 @@ public enum KeyboardType: Codable, Equatable, Identifiable {
      `.email` represents keyboards that have e-mail specific
      input keys for the current locale.
      
-     KeyboardKit currently has no built-in views that can be
-     used to implement this kind of keyboard. If you want to
-     use it, you must create your own keyboard view.
+     KeyboardKit has no built-in view for this keyboard type,
+     so if you want to use it, you must create your own.
      */
     case email
     
@@ -58,9 +57,9 @@ public enum KeyboardType: Codable, Equatable, Identifiable {
      `.emoji` represents keyboards that either present emoji
      characters or emoji categories.
      
-     This can be implemented with the `EmojiKeyboard` or the
-     `EmojiCategoryKeyboard` but you can create custom emoji
-     keyboards as well.
+     This type can be rendered with the ``EmojiKeyboard`` or
+     the ``EmojiCategoryKeyboard`` views, but you can create
+     custom emoji keyboards as well.
      */
     case emojis
     
@@ -68,9 +67,8 @@ public enum KeyboardType: Codable, Equatable, Identifiable {
      `.image` represents keyboards that present custom image
      buttons with custom handing.
      
-     KeyboardKit currently has no built-in views that can be
-     used to implement this kind of keyboard. If you want to
-     use it, you must create your own keyboard view.
+     KeyboardKit has no built-in view for this keyboard type,
+     so if you want to use it, you must create your own.
      */
     case images